security-audit/fixed-code/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
Flask-WTF==1.2.1
python-dotenv==1.0.0
bleach==6.1.0
